requirements.txt
# pytorch3d-0.7.7
torch==2.2.0 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.17.0 --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.2.0 --index-url https://download.pytorch.org/whl/cu118
lightning==2.2.0
roma==1.5.3
imageio[pyav]
imageio[ffmpeg]
lmdb==1.6.2
open3d=0.19.0
plyfile==1.0.3
omegaconf==2.3.0
colored==2.3.0
rich==14.0.0
opencv-python==4.11.0.86
chumpy==0.70
xformers==0.0.24 --index-url https://download.pytorch.org/whl/cu118
gradio
tyro==0.8.0
onnxruntime-gpu==1.18
onnx==1.16
mediapipe==0.10.21
easydict==1.13
kornia==0.1.9
transformers==4.57.0
configer==1.3.1
torchgeometry==0.1.2
pynvml==13.0.1
numpy==1.26.4